refactor: separate unrelated helpers, renamed app instance helper to better express intent
