+++ test/e2e/helpers/usernameHelper.js
@@ -1,0 +1,10 @@
+const usernameChars = '0123456789abcdefghijklmnopqrstuvwxyz';
+const getRandomUsername = () => {
+    let username = '';
+    for (let i = 0; i < 30; i++) {
+        username += usernameChars[Math.floor(Math.random() * usernameChars.length)];
+    }
+    return username;
+};
+
+module.exports = { getRandomUsername };
