+++ test/e2e/rooms/supportCode/rooms.js
@@ -1,5 +1,5 @@
 const defineSupportCode = require('cucumber').defineSupportCode;
-const getNewAppInstance = require('../../config');
+const getAppInstance = require('../../helpers/appConfig');
 const { when } = require('mobx');
 
 defineSupportCode(({ Before, Given, Then, When }) => {
@@ -10,7 +10,7 @@ defineSupportCode(({ Before, Given, Then, When }) => {
     let chat;
 
     Before((testCase, done) => {
-        app = getNewAppInstance();
+        app = getAppInstance();
         when(() => app.socket.connected, done);
     });
